backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-multipart==0.0.12
pydantic==2.9.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
pandas==2.2.2
openpyxl==3.1.5
python-docx==1.1.2
aiofiles==24.1.0
httpx==0.27.2
numpy==1.26.4
